PB-368 : move all style utils into relevant file
some marker style could be re-used with things that are not marker...
