+++ src/scss/variables-admin.module.scss
@@ -56,6 +56,8 @@ $mocassin-to-red-2: #ff7e29;
 
 :export {
   white: #{$white};
+  black: #{$black};
+  malibu: #{$malibu};
   red: #{$red};
   mocassin: #{$mocassin};
   mocassinToRed1: #{$mocassin-to-red-1};
